requirements.txt
fastapi>=0.124,<0.125
httpx>=0.28,<0.29
psycopg2-binary>=2.9,<2.10
pydantic>=2.12,<2.13
pydantic-settings>=2.12,<2.13
python-dotenv>=1.2,<1.3
pytest>=9.0,<9.1
sqlmodel==0.0.27
uvicorn>=0.38,<0.39
watchfiles>=1.1,<1.2
